Explicitly set an app menu when not on OS X
Sadly, the prefer-app-menu heuristics don't quite work
right in all cases for windows with headerbars, which
never display a menubar. There are cases where the heuristics
expect a menubar to be displayed inside the window, and
don't provide an app menu. But with our header bar, we
always want our app menu, _except_ for the OS X case,
where we want our menubar with the platform-provided
app mnenu.